requirements.txt
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0
xgboost==2.0.3
sentence-transformers==2.7.0
matplotlib==3.8.2
seaborn==0.13.2
plotly==5.20.0
streamlit==1.31.0
google-auth==2.28.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.118.0
python-dotenv==1.0.1
joblib==1.3.2
nltk==3.8.1